Move GroupKey -> Value converting codes from hash.rs to value.rs
