+++ src/executor/mod.rs
@@ -10,7 +10,7 @@ mod limit;
 mod select;
 mod update;
 
-pub use aggregate::AggregateError;
+pub use aggregate::{AggregateError, GroupKey};
 pub use blend::BlendError;
 pub use context::{BlendContextError, FilterContextError, UnionContextError};
 pub use evaluate::EvaluateError;
